sgp_dd: Rework Android support
Explicit platform type tests are a nightmare from the point of view
of maintaining cross-platform software. Hence rework Android support
using primitives that are available on all supported platforms.

Note: The autoconf approach to portability is to test for features,
not for platforms or versions. Testing for platforms or versions
goes against the autoconf philosophy.

Signed-off-by: Bart Van Assche <bart.vanassche@wdc.com>